Added Readme.txt
Find the changes.txt file!

The first commit here adds the readme.txt file that explains the purpose
of the whole exercise.  GitHub requests adding a file with this name as
an overview of each repository.  There is also the default .gitignore
file with tells the git source control to ignore backup files and the
.gitattributes for rules about merging project files.   The .gitignore
and .gitattributes are just the default files generated by github for
every new repository.