Adding the game.html and game.js files
Added game.html as the HTML5 page that calls your javascript game named
game.js.  In game.js, you will write your game but it fails in this
version because of a missing reference 'foobar'.
This would be a good time to learn how to launch the browser from your
editor and to find the Javascript debugger from your browser.